requirements.txt
web3>=6.11.0
pyyaml>=6.0
requests>=2.31.0
python-dotenv>=1.0.0
